+++ src/raygun.js
@@ -39,7 +39,7 @@ var raygunFactory = function(window, $, undefined) {
     _disableErrorTracking = false,
     _disablePulse = true,
     _wrapAsynchronousCallbacks = false,
-    _trackMeasurementTimings = false,
+    _automaticPerformanceCustomTimings = false,
     _customData = {},
     _tags = [],
     _user,
@@ -119,7 +119,7 @@ var raygunFactory = function(window, $, undefined) {
         _pulseCustomLoadTimeEnabled = options.pulseCustomLoadTimeEnabled || false;
         _setCookieAsSecure = options.setCookieAsSecure || false;
         _captureMissingRequests = options.captureMissingRequests || false;
-        _trackMeasurementTimings = options.trackMeasurementTimings || false;
+        _automaticPerformanceCustomTimings = options.automaticPerformanceCustomTimings || false;
 
         if (options.apiUrl) {
           _raygunApiUrl = options.apiUrl;
@@ -500,7 +500,7 @@ var raygunFactory = function(window, $, undefined) {
           _beforeSendRumCallback,
           _setCookieAsSecure,
           _captureMissingRequests,
-          _trackMeasurementTimings
+          _automaticPerformanceCustomTimings
         );
         _rum.attach();
       };